Split EditCloudSettings tabs into separate files
